requirements.txt
ansicon==1.89.0
arrow==1.3.0
attrs==24.2.0
binaryornot==0.4.4
blessed==1.20.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cookiecutter==2.6.0
cycler==0.12.1
distlib==0.3.9
editor==1.6.6
filelock==3.16.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
fonttools==4.55.2
greenlet==3.3.0
h11==0.14.0
idna==3.10
inquirer==3.4.0
inquirerpy==0.3.4
itsdangerous==2.2.0
Jinja2==3.1.4
jinxed==1.3.0
joblib==1.4.2
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
mdurl==0.1.2
MouseInfo==0.1.3
numpy==2.1.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pfzy==0.3.4
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
PyMySQL==1.1.2
pyparsing==3.2.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySide6==6.9.0
PySide6_Addons==6.9.0
PySide6_Essentials==6.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytweening==1.2.0
pytz==2024.2
PyYAML==6.0.2
readchar==4.2.1
requests==2.32.3
rich==13.9.4
runs==1.2.2
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.26.1
setuptools==75.3.0
shiboken6==6.9.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.45
text-unidecode==1.3
threadpoolctl==3.5.0
trio==0.27.0
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20241003
types-pytz==2024.2.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.27.1
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.1
wsproto==1.2.0
xmod==1.8.1
